requirements.txt
Django==2.2.10
djangorestframework==3.12.4
drf-nested-routers==0.93.4
Markdown==3.3.4
pytz==2021.3
sqlparse==0.4.2
